Move `function_call_parser.py` to function_call module
- `function_call_parser.py` has over 1000 lines. It will become more and more difficult to maintain
